requirements.txt
numpy 
pandas
matplotlib
notebook
PyYAML
Flask
from_root
ensure
python-box
tqdm
joblib
gdown
keras
tensorflow
-e .